Refactor pr#36492 code style
