+++ database/connector/core/src/main/java/org/apache/shardingsphere/database/connector/core/type/DatabaseTypeFactory.java
@@ -22,7 +22,10 @@ import lombok.NoArgsConstructor;
 import org.apache.shardingsphere.database.connector.core.exception.UnsupportedStorageTypeException;
 import org.apache.shardingsphere.infra.exception.ShardingSpherePreconditions;
 import org.apache.shardingsphere.infra.spi.ShardingSphereServiceLoader;
+import org.apache.shardingsphere.infra.spi.type.typed.TypedSPILoader;
 
+import java.sql.DatabaseMetaData;
+import java.sql.SQLException;
 import java.util.Collection;
 import java.util.stream.Collectors;
 
@@ -47,6 +50,17 @@ public final class DatabaseTypeFactory {
             }
         }
         return databaseTypes.iterator().next();
+    }
+    
+    /**
+     * Get database type.
+     *
+     * @param metaData database meta data
+     * @return database type
+     * @throws SQLException SQL exception
+     */
+    public static DatabaseType get(final DatabaseMetaData metaData) throws SQLException {
+        return metaData.getDatabaseProductName().contains("Hive") ? TypedSPILoader.getService(DatabaseType.class, "Hive") : get(metaData.getURL());
     }
     
     private static boolean matchURLs(final String url, final DatabaseType databaseType) {
