+++ test/e2e/sql/src/test/java/org/apache/shardingsphere/test/e2e/sql/it/sql/dml/BaseDMLE2EIT.java
@@ -35,10 +35,10 @@ import org.apache.shardingsphere.test.e2e.sql.cases.dataset.DataSetLoader;
 import org.apache.shardingsphere.test.e2e.sql.cases.dataset.metadata.DataSetColumn;
 import org.apache.shardingsphere.test.e2e.sql.cases.dataset.metadata.DataSetMetaData;
 import org.apache.shardingsphere.test.e2e.sql.cases.dataset.row.DataSetRow;
-import org.apache.shardingsphere.test.e2e.sql.framework.metadata.DialectQueryBehaviorProvider;
 import org.apache.shardingsphere.test.e2e.sql.env.DataSetEnvironmentManager;
 import org.apache.shardingsphere.test.e2e.sql.env.SQLE2EEnvironmentEngine;
 import org.apache.shardingsphere.test.e2e.sql.framework.metadata.DialectDatabaseAssertionMetaDataFactory;
+import org.apache.shardingsphere.test.e2e.sql.framework.metadata.DialectQueryBehaviorProvider;
 import org.apache.shardingsphere.test.e2e.sql.framework.param.model.AssertionTestParameter;
 import org.apache.shardingsphere.test.e2e.sql.framework.param.model.CaseTestParameter;
 import org.apache.shardingsphere.test.e2e.sql.framework.param.model.E2ETestParameter;
@@ -235,27 +235,17 @@ public abstract class BaseDMLE2EIT implements SQLE2EIT {
     private String generateFetchActualDataSQL(final Map<String, DataSource> actualDataSourceMap, final DataNode dataNode, final DatabaseType databaseType) throws SQLException {
         String tableName = dataNode.getTableName();
         Optional<String> primaryKeyColumnName = DialectDatabaseAssertionMetaDataFactory.getPrimaryKeyColumnName(databaseType, actualDataSourceMap.get(dataNode.getDataSourceName()), tableName);
-        if (primaryKeyColumnName.isPresent()) {
-            return String.format("SELECT * FROM %s ORDER BY %s ASC", tableName, primaryKeyColumnName.get());
-        }
-        Optional<DialectQueryBehaviorProvider> behaviorProvider = DatabaseTypedSPILoader.findService(DialectQueryBehaviorProvider.class, databaseType);
-        if (behaviorProvider.isPresent()) {
-            Optional<String> fallbackOrderBy = behaviorProvider.get().getFallbackOrderByWhenNoPrimaryKey();
-            if (fallbackOrderBy.isPresent()) {
-                return String.format("SELECT * FROM %s ORDER BY %s", tableName, fallbackOrderBy.get());
-            }
-        }
-        return String.format("SELECT * FROM %s", tableName);
+        return primaryKeyColumnName.map(optional -> String.format("SELECT * FROM %s ORDER BY %s ASC", tableName, optional))
+                .orElseGet(() -> DatabaseTypedSPILoader.findService(DialectQueryBehaviorProvider.class, databaseType)
+                        .flatMap(DialectQueryBehaviorProvider::getFallbackOrderByWhenNoPrimaryKey).map(optional -> String.format("SELECT * FROM %s ORDER BY %s", tableName, optional))
+                        .orElseGet(() -> String.format("SELECT * FROM %s", tableName)));
     }
     
     private void assertMetaData(final ResultSetMetaData actual, final Collection<DataSetColumn> expected) throws SQLException {
         assertThat(actual.getColumnCount(), is(expected.size()));
         int index = 1;
         for (DataSetColumn each : expected) {
-            String actualLabel = actual.getColumnLabel(index++);
-            int lastDotIndex = actualLabel != null ? actualLabel.lastIndexOf('.') : -1;
-            String normalizedLabel = lastDotIndex >= 0 ? actualLabel.substring(lastDotIndex + 1) : actualLabel;
-            assertThat(normalizedLabel.toUpperCase(), is(each.getName().toUpperCase()));
+            assertThat(actual.getColumnLabel(index++).toUpperCase(), is(each.getName().toUpperCase()));
         }
     }
