+++ test/e2e/sql/src/test/resources/cases/dml/dataset/encrypt/batch_insert_into_user_table_for_hive.xml
@@ -15,16 +15,16 @@
   ~ limitations under the License.
   -->
 
-<dataset update-count="1">
+<dataset update-count="2">
     <metadata data-nodes="encrypt.t_user">
-        <column name="user_id" type="numeric" />
-        <column name="user_name_cipher" type="varchar" />
-        <column name="user_name_like" type="varchar" />
-        <column name="password_cipher" type="varchar" />
-        <column name="email_cipher" type="varchar" />
-        <column name="user_telephone_cipher" type="varchar" />
-        <column name="user_telephone_like" type="varchar" />
-        <column name="creation_date" type="datetime" />
+        <column name="t_user.user_id" type="numeric" />
+        <column name="t_user.user_name_cipher" type="varchar" />
+        <column name="t_user.user_name_like" type="varchar" />
+        <column name="t_user.password_cipher" type="varchar" />
+        <column name="t_user.email_cipher" type="varchar" />
+        <column name="t_user.user_telephone_cipher" type="varchar" />
+        <column name="t_user.user_telephone_like" type="varchar" />
+        <column name="t_user.creation_date" type="datetime" />
     </metadata>
     <row data-node="encrypt.t_user" values="10, sVq8Lmm+j6bZE5EKSilJEQ==, yi`mht`m, aQol0b6th65d0aXe+zFPsQ==, WM0fHOH91JNWnHTkiqBdyNmzk4uJ7CCz4mB1va9Ya1M=, kLjLJIMnfyHT2nA+viaoaQ==, 01454589811, 2017-08-08" />
     <row data-node="encrypt.t_user" values="11, fQ7IzBxKVuNHtUF6h6WSBg==, mhth, wuhmEKgdgrWQYt+Ev0hgGA==, svATu3uWv9KfiloWJeWx3A==, 0kDFxndQdzauFwL/wyCsNQ==, 01454589810, 2017-08-08" />
@@ -56,4 +56,6 @@
     <row data-node="encrypt.t_user" values="37, aXS0VfnqHIAnOAtDjsF/9Q==, 伈嶱啴, bO/8ha1eS/H8/3DugjdOAQ==, fwyOxfHtLxNuSCFmghYiY0qMsgbpjg5UIo3xmJOLGu0=, 60fpnMdKCWeyKzxkdthn2Q==, 09101401454, 2017-08-08" />
     <row data-node="encrypt.t_user" values="38, 59/68izQEdnwNSueX1lPAA==, 伈乄, ilD/Tk7DUG4+EuznS1bNLg==, 2emhAeiXPr0kHbFrhYlM1dmzk4uJ7CCz4mB1va9Ya1M=, 60fpnMdKCWeyKzxkdthn2Q==, 09101401454, 2017-08-08" />
     <row data-node="encrypt.t_user" values="39, fn9LnNltUAOWO0F0iy0+Jw==, 伈妅, qe/WdUiSPP1RAsSSuejGJw==, zx2omwIbXHpEJeh8ta7HqQq2ZLhWcqfQ8/EQnIqMx+g=, 60fpnMdKCWeyKzxkdthn2Q==, 09101401454, 2017-08-08" />
+    <row data-node="encrypt.t_user" values="40, Dit2zNmv7pj+Jf5ND1tGQQ==, upl`t, MyOShk4kjRnds7CZfU5NCw==, VvWKtMDr5+58OMAAu6BH5A==, 8Rj1Bg696vJMKwPNxzxGOg==, 01450145014, 2018-08-08" />
+    <row data-node="encrypt.t_user" values="41, YTfHI/5xnPNvpjCR20sPWw==, lhld, qbUuBn0oxdrV8sNNyoqDCg==, FRo1h44oclER1+0MeI1T0w==, cD36lxxtVApK4QLzec87zg==, 14541454545, 2019-08-08" />
 </dataset>
