+++ test/e2e/sql/src/test/resources/cases/dml/e2e-dml-insert.xml
@@ -256,6 +256,6 @@
     </test-case>
 
     <test-case sql="INSERT INTO t_user (user_id, user_name, password, email, telephone, creation_date) VALUES (?, ?, '123456', 'tomas@gmail.com', '12341234123', '2018-08-08'), (?, ?, '23456', 'mike@qq.com', '23452345456', '2019-08-08')" db-types="Hive" scenario-types="encrypt" skip-batch="true">
-        <assertion parameters="40:int, tomas:String, 41:int, mike:String" expected-data-file="batch_insert_into_user_table.xml" />
+        <assertion parameters="40:int, tomas:String, 41:int, mike:String" expected-data-file="batch_insert_into_user_table_for_hive.xml" />
     </test-case>
 </e2e-test-cases>
